Update LastAppliedTopology to TopologyRef interface
TopologyRef is able to reference Name and Namespace but without this
patch only TopologyRef.Name is saved to .Status.LastAppliedTopology.

This patch aligns both interfaces to use the same type in the same form.
In addition, some logic is moved to infra-operator because it is common
to all operators.

It also updates the SpecCore webhook interface to accept Namespace as
parameter. It ensures that the referenced topology from
openstack-operator
is properly validated and rejects requests with a different namespace.

This patch is based on [1] by fpantano@redhat.com

Depends-On: openstack-k8s-operators/openstack-operator#1306

[1] https://github.com/openstack-k8s-operators/manila-operator/pull/390

Signed-off-by: John Fulton <fulton@redhat.com>